+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Types of changes
 ## [1.21.0]
 
 - `Added` new features `findInCSV` mask to get one or multiple csv lines which matched with Json entry value from CSV files.
+- `Added` new flag `--serve` to start a HTTP server that apply masking to JSON body of requests.
 
 ## [1.20.0]